+++ src/services/query/index.ts
@@ -1,0 +1,1 @@
+export * from "./ReactQueryProvider";
